Chapter-18/image_gen_function/requirements.txt
functions-framework==3.*
vertexai
Pillow
google-cloud-storage
